requirements.txt
customtkinter==5.2.2
Pillow==12.1.1
plyer==2.1.0
yt_dlp==2026.2.21
